refactor(api): implement validated command->apply->ack->state flow.
Inject dependencies into commands via CommandsContext, remove extern coupling, and dispatch commands through per-type handlers.
Apply callbacks now persist to NVS, update applied config, and return result for ACK.
MQTT transport forwards cmd payloads, publishes ACK/state from main loop, and retains canonical state snapshots.

+++ esp32/level_sensor/device_state.h
@@ -61,15 +61,15 @@ struct ProbeInfo
     bool connected;
     ProbeQualityReason quality;
     SenseMode senseMode;
-    uint16_t raw;
+    int32_t raw; // 32-bit raw reading for consistency across probe/calibration paths
     bool rawValid;
 };
 
 struct CalibrationInfo
 {
     CalibrationState state;
-    uint16_t dry;
-    uint16_t wet;
+    int32_t dry;
+    int32_t wet;
     bool inverted;
     uint16_t minDiff;
 };
